Returns time series of Daily Maximum and Minimum with a random multi-realization obtained by using newVARmultieventRealization. This function is called by ComprehensiveTemperatureGenerator.

Returns time series of Daily Maximum and Minimum with a random multi-realization obtained by using newVARmultieventRealization. This function is called by ComprehensiveTemperatureGenerator.

generateTemperatureTimeseries(
  std_tn,
  std_tx,
  SplineTx,
  SplineTn,
  SplineTm,
  SplineDeltaT,
  std_tm,
  var = NULL,
  exogen = NULL,
  normalize = TRUE,
  type = 3,
  extremes = TRUE,
  sample = NULL,
  option = 1,
  original_data,
  origin_x = NULL,
  origin_data = NULL,
  noise = NULL
)

Arguments

std_tn

vector containing standard deviation of daily minimum temperature anomalies. stdTn is default, see setComprehensiveTemperatureGeneratorParameters.

std_tx

vector containing standard deviation of daily maximum temperature anomalies. stdTx is default, see setComprehensiveTemperatureGeneratorParameters.

SplineTx

matrix containing the averaged daily maximum temperature obtained by a spline interpolation of monthly means . SplineAdvTx is default, see setComprehensiveTemperatureGeneratorParameters.

SplineTn

matrix containing the averaged daily minimum temperature obtained by a spline interpolation of monthly means . SplineAdvTn is default, see setComprehensiveTemperatureGeneratorParameters.

SplineTm

matrix containing the averaged daily "mean" temperature obtained by a spline interpolation of monthly means . SplineAdvTm is default, see setComprehensiveTemperatureGeneratorParameters.

SplineDeltaT

matrix containing the rescaled averaged daily temperature range obtained by a spline interpolation of monthly means. SplineAdvDelta_T_sim/SplineAdvDelta_T is default, see setComprehensiveTemperatureGeneratorParameters.

std_tm

vector containing standard deviation of daily "mean" temperature anomalies. stdTn is default, see setComprehensiveTemperatureGeneratorParameters.

var

A VAR model represented by a varest object as returned by getVARmodel or VAR

exogen

see VAR

normalize

logical variable If TRUE normalizeGaussian_severalstations is used, otherwise not. If option is 2, it is always TRUE.

type

see quantile

sample, origin_x, origin_data, extremes

see normalizeGaussian_severalstations

option

integer value. If 1, the generator works with minimum and maximum temperature, if 2 (Default) it works with th average value between maximum and minimum temparature and the respective daily Thermal Range.

original_data

matrix containing the measured standardized temperature anomalies

noise

stochastic noise to add for variabile generation. Default is NULL. See newVARmultieventRealization.

Value

This function returns a list of the following variables:

res_multigen matrix containing standardized values of daily maximum and minimum temperature anomalies

Tx_spline matrix containing climatic "spline-interpolated" daily maximum temperature

Tn_spine matrix containing climatic "spline-interpolated" daily minimum temperature

Tx_gen matrix containing generated daily maximum daily temperature (\(Tx_{gen}\))

Tn_gen matrix containing generated daily minimum daily temperature (\(Tn_{gen}\))

Tm_gen matrix containing generated "mean" daily temperature defined as \(\frac{Tx_{gen}+Tn_{gen}}{2}\)

DeltaT_gen matrix containing generated daily thermal range defined as \(Tx_{gen}-Tn_{gen}\)

See the R code for further details

See also

Author

Emanuele Cordano, Emanuele Eccel