Calculates the daily means of a range of days around each date of a data frame corresponding to a period between year_min and year_max for stations listed in station

Calculates the daily means of a range of days around each date of a data frame corresponding to a period between year_min and year_max for stations listed in station

getDailyMean(
  data,
  year_min = 1961,
  year_max = 1990,
  station = c("T0001", "T0010"),
  origin = "1961-1-1",
  lag = 5
)

Arguments

data

a data frame containing daily data.

year_min

start year

year_max

end year

station

character vector of the IDs of the station where the data are requested

origin

origin date of time-series

lag

lag (number of days) on which daily mean is calculated. The mean is calculated considereing lag days before and after each day.

Value

a matrix containing the requested daily mean data where each day corresponds to a row and each station corresponds to a column

Note

The input data frame data must have the following fields: year,month,day,variables_ID1,variables_ID2,... where the fields ,variables_ID1,variables_ID2,... contain the daily variables referred to the respective stations and the field names are replaced with the respective station ID.

See also

Author

Emanuele Cordano, Emanuele Eccel