Generates several realizations of a VAR model

Generates several realizations of a VAR model

newVARmultieventRealization(
  var,
  xprev = rnorm(var@VAR$K * var@VAR$p),
  exogen = NULL,
  nrealization = 10,
  B = t(chol(cov(residuals(var)))),
  extremes = TRUE,
  type = 3,
  noise = NULL
)

Arguments

var

A VAR model represented by a varest2 object as returned by getVARmodel

xprev

previous status of the random variable

exogen

matrix containing the values of the "exogen" variables (predictor) for the generation

nrealization

number of realization (e.g. days to simulate). If exogen is not NULL and it is a matrix, it must be lower or equal to the number of rows of exogen

B

matrix of coefficients for the vector white-noise component

extremes, type

see inv_GPCA

noise

stochastic noise to add for variabile generation. Default is NULL and it is automatically randomly genereted accordind to matrix B. If the VAR model (var argument) does not fit well the residuals (e.g. non-normality, non-serialty or heteroskesticity) and the white noise is manually inserted, in this case argument B is not taken into account.

Value

a matrix of values

Author

Emanuele Cordano, Emanuele Eccel