Converts a random variable x extracted by a population represented by the sample data or sample to a normally-distributed variable with assigned mean and standard deviation or vice versa in case inverse is TRUE

Converts a random variable x extracted by a population represented by the sample data or sample to a normally-distributed variable with assigned mean and standard deviation or vice versa in case inverse is TRUE

normalizeGaussian(
  x = 0,
  data = x,
  cpf = NULL,
  mean = 0,
  sd = 1,
  inverse = FALSE,
  step = NULL,
  prec = 10^-4,
  type = 3,
  extremes = TRUE,
  sample = NULL
)

Arguments

x

value or vector of values to be converted

data

a sample of data on which a non-parametric probability distribution is estimated

cpf

cumulative probability distribution. If NULL (default) is calculated as ecdf(data)

mean

mean (expected value) of the normalized random variable. Default is 0.

sd

standard deviation of the normalized random variable. Default is 1.

inverse

logical value. If TRUE the function works inversely (the opposite way). Default is FALSE.

step

vector of values in which step discontinuities of the cumulative probability function occur. Default is NULL

prec

amplitude of the neighbourhood of the step discontinuities where cumulative probability function is treated as non-continuous.

type

see quantile

extremes

logical variable. If TRUE (default) the probability or frequency is multiplied by $$\frac{N}{N+1}$$ where \(N\) is the length of data

sample

a character string or NULL containing sample or probability distribution information. Default is NULL

Value

the normalized variable or its inverse

@note This function makes a Marginal Gaussianization. See the R code for further details

Author

Emanuele Cordano, Emanuele Eccel