It makes the Q-Q plots observed vs generated time series of daily maximum, minimum temperature and daily thermal range for a list of collected stochastic generations

It makes the Q-Q plots observed vs generated time series of daily maximum, minimum temperature and daily thermal range for a list of collected stochastic generations

qqplot_RMAWGEN_Tx(
  Tx_mes,
  Tx_gen,
  Tn_gen,
  Tn_mes,
  Tx_spline = NULL,
  Tn_spline = NULL,
  xlab = "observed",
  ylab = "simulated",
  when = 1:nrow(Tx_mes),
  main = names(Tx_gen),
  station,
  pdf = NULL,
  xlim = range(Tx_mes),
  ylim = xlim,
  cex = 0.4,
  cex.main = 1,
  cex.lab = 1,
  cex.axis = 1
)

qqplot_RMAWGEN_Tn(
  Tx_mes,
  Tx_gen,
  Tn_gen,
  Tn_mes,
  Tx_spline = NULL,
  Tn_spline = NULL,
  xlab = "observed",
  ylab = "simulated",
  when = 1:nrow(Tn_mes),
  main = names(Tn_gen),
  station,
  pdf = NULL,
  xlim = range(Tn_mes),
  ylim = xlim,
  cex = 0.4,
  cex.main = 1,
  cex.lab = 1,
  cex.axis = 1
)

qqplot_RMAWGEN_deltaT(
  Tx_mes,
  Tx_gen,
  Tn_gen,
  Tn_mes,
  xlab = "observed",
  ylab = "simulated",
  when = 1:nrow(Tx_mes),
  main = names(Tx_gen),
  station,
  pdf = NULL,
  xlim = range(Tx_mes - Tn_mes),
  ylim = xlim,
  cex = 0.4,
  cex.main = 1,
  cex.lab = 1,
  cex.axis = 1
)

qqplot_RMAWGEN_prec(
  prec_mes,
  prec_gen,
  xlab = "observed",
  ylab = "simulated",
  when = 1:nrow(prec_mes),
  main = names(prec_gen),
  station,
  pdf = NULL,
  xlim = range(prec_mes),
  ylim = xlim,
  cex = 0.4,
  cex.main = 1,
  cex.lab = 1,
  cex.axis = 1,
  lag = 1
)

Arguments

Tx_mes

data frame containing measured daily maximum temperature

Tx_gen

data frame containing generated daily maximum temperature

Tn_gen

data frame containing generated daily minimum temperature

Tn_mes

data frame containing measured daily minimum temperature

Tx_spline

data frame containing spline-interpolated daily maximum temperature. Default is NULL and not considered for Q-Q plot.

Tn_spline

data frame containing spline-interpolated daily minimum temperature Default is NULL and not considered for Q-Q plot.

xlab, ylab

lables of x and y axes. See qqplot.

when

day indices on which the data frame are extracted for Q-Q plot. Default is 1:nrow(Tn_mes) (in qqplot_RMAWGEN_Tn) or 1:nrow(Tx_mes) (otherwise)

main

main titles for each plot. Default is names(Tn_gen) (in qqplot_RMAWGEN_Tn) or names(Tx_gen) (otherwise)

station

identification name (ID) of the station used for the Q-Q plot

pdf

name of pdf file if output is written in a pdf file

xlim

see qqplot. Default is range(Tn_mes) (in qqplot_RMAWGEN_Tn) or range(Tx_mes) (in qqplot_RMAWGEN_Tx) .or range(Tx_mes-Tn_mes) (in qqplot_RMAWGEN_deltaT)

ylim, cex, cex.main, cex.lab, cex.axis

see qqplot and plot

prec_mes

data frame containing measured daily precipitation (in millimeters)

prec_gen

data frame containing generated daily precipitation (in millimeters)

lag

lag (current index included) on whose value the precipitation addition is made. See qqplot.lagged.

Note

Tx_gen,Tn_gen and main must have an even number of elements.

Author

Emanuele Cordano